IGNITE-17590 Re-factoring
