+++ modules/platforms/cpp/ignite/client/detail/table/table_impl.h
@@ -18,6 +18,7 @@
 #pragma once
 
 #include "ignite/client/detail/cluster_connection.h"
+#include "ignite/client/detail/table/schema.h"
 #include "ignite/common/uuid.h"
 
 #include <memory>
@@ -25,157 +26,6 @@
 #include <unordered_map>
 
 namespace ignite::detail {
-
-// TODO: Move to separate file
-/**
- * Client data types.
- */
-enum class client_data_type
-{
-    /// Byte.
-    INT8 = 1,
-
-    /// Short.
-    INT16 = 2,
-
-    /// Int.
-    INT32 = 3,
-
-    /// Long.
-    INT64 = 4,
-
-    /// Float.
-    FLOAT = 5,
-
-    /// Double.
-    DOUBLE = 6,
-
-    /// Decimal.
-    DECIMAL = 7,
-
-    /// UUID / Guid.
-    UUID = 8,
-
-    /// String.
-    STRING = 9,
-
-    /// Byte array.
-    BYTES = 10,
-
-    /// BitMask.
-    BITMASK = 11,
-
-    /// Local date.
-    DATE = 12,
-
-    /// Local time.
-    TIME = 13,
-
-    /// Local date and time.
-    DATE_TIME = 14,
-
-    /// Timestamp (instant).
-    TIMESTAMP = 15,
-
-    /// Number (BigInt).
-    NUMBER = 16,
-};
-
-/**
- * Get client data type from int value.
- *
- * @param val Value.
- * @return Matching client data type.
- */
-inline client_data_type client_data_type_from_int(std::int32_t val) {
-    if (val > std::int32_t(client_data_type::NUMBER) || val < std::int32_t(client_data_type::INT8))
-        throw ignite_error("Value is out of range for client data type: " + std::to_string(val));
-    return client_data_type(val);
-}
-
-/**
- * Column.
- */
-struct column {
-    std::string name;
-    client_data_type type{};
-    bool nullable{false};
-    bool is_key{false};
-    std::int32_t schema_index{0};
-    std::int32_t scale{0};
-
-    /**
-     * Unpack column from MsgPack object.
-     *
-     * @param object MsgPack object.
-     * @return Column value.
-     */
-    [[nodiscard]] static column unpack(const msgpack_object& object) {
-        if (object.type != MSGPACK_OBJECT_ARRAY)
-            throw ignite_error("Schema column expected to be serialized as array");
-
-        const msgpack_object_array& arr = object.via.array;
-
-        constexpr std::uint32_t expectedCount = 6;
-        assert(arr.size >= expectedCount);
-
-        column res{};
-        res.name = protocol::unpack_object<std::string>(arr.ptr[0]);
-        res.type = client_data_type_from_int(protocol::unpack_object<std::int32_t>(arr.ptr[1]));
-        res.is_key = protocol::unpack_object<bool>(arr.ptr[2]);
-        res.nullable = protocol::unpack_object<bool>(arr.ptr[3]);
-        res.scale = protocol::unpack_object<std::int32_t>(arr.ptr[5]);
-
-        return std::move(res);
-    }
-};
-
-/**
- * Schema.
- */
-struct schema {
-    std::int32_t version{-1};
-    std::int32_t key_column_count{0};
-    std::vector<column> columns;
-
-    // Default
-    schema() = default;
-
-    /**
-     * Constructor.
-     *
-     * @param version Version.
-     * @param key_column_count Key column count.
-     * @param columns Columns.
-     */
-    schema(std::int32_t version, std::int32_t key_column_count, std::vector<column>&& columns)
-        : version(version)
-        , key_column_count(key_column_count)
-        , columns(std::move(columns)) {}
-
-    /**
-     * Read schema using reader.
-     *
-     * @param reader Reader to use.
-     * @return Schema instance.
-     */
-    static std::shared_ptr<schema> read(const msgpack_object_kv& object) {
-        auto schema_version = protocol::unpack_object<std::int32_t>(object.key);
-        std::int32_t key_column_count = 0;
-
-        std::vector<column> columns;
-        columns.reserve(protocol::unpack_array_size(object.val));
-
-        protocol::unpack_array_raw(object.val, [&columns, &key_column_count](const msgpack_object& object) {
-            auto val = column::unpack(object);
-            if (val.is_key)
-                ++key_column_count;
-            columns.emplace_back(std::move(val));
-        });
-
-        return std::make_shared<schema>(schema_version, key_column_count, std::move(columns));
-    }
-};
 
 /**
  * Table view implementation.
@@ -212,21 +62,7 @@ public:
      *
      * @return Latest schema.
      */
-    void get_latest_schema_async(const ignite_callback<std::shared_ptr<schema>>& callback) {
-        auto latest_schema_version = m_latest_schema_version;
-
-        if (latest_schema_version >= 0) {
-            std::shared_ptr<schema> schema;
-            {
-                std::lock_guard<std::mutex> guard(m_schemas_mutex);
-                schema = m_schemas[latest_schema_version];
-            }
-            callback({std::move(schema)});
-            return;
-        }
-
-        load_schema_async(callback);
-    }
+    void get_latest_schema_async(const ignite_callback<std::shared_ptr<schema>>& callback);
 
 private:
     /**
@@ -234,30 +70,7 @@ private:
      *
      * @return Latest schema.
      */
-    void load_schema_async(ignite_callback<std::shared_ptr<schema>> callback) {
-        auto writer_func = [&](protocol::writer &writer) {
-            writer.write(m_id);
-            writer.write_nil();
-        };
-
-        auto table = shared_from_this();
-        auto reader_func = [table](protocol::reader &reader) mutable -> std::shared_ptr<schema> {
-            auto schema_cnt = reader.read_array_size();
-            if (!schema_cnt)
-                throw ignite_error("Schema not found");
-
-            std::shared_ptr<schema> last;
-            reader.read_map_raw([&last, &table](const msgpack_object_kv& object) {
-                last = schema::read(object);
-                table->add_schema(last);
-            });
-
-            return std::move(last);
-        };
-
-        m_connection->perform_request<std::shared_ptr<schema>>(
-            client_operation::TABLE_GET, writer_func, std::move(reader_func), std::move(callback));
-    }
+    void load_schema_async(ignite_callback<std::shared_ptr<schema>> callback);
 
     /**
      * Add schema.
